backend/requirements.txt
Django
djangorestframework
pandas
matplotlib
reportlab
